requirements.txt
djangorestframework
coreapi